+++ index.js
@@ -16,12 +16,22 @@ import { getHelpersByPlatform } from "./src/platforms.js";
 import { checkedToString, toArrayIfNecessary } from "./src/reflection.js";
 
 /**
+ * Build error messages for unsupported shells.
+ *
+ * @param {string} shellName The full name of a shell.
+ * @returns {string} The unsupported shell error message.
+ */
+function shellError(shellName) {
+  return `Shescape does not support the shell ${shellName}`;
+}
+
+/**
  * A class to escape user-controlled inputs to shell commands to prevent shell
  * injection.
  *
  * @example
  * import { spawn } from "node:child_process";
- * const shescape = Shescape();
+ * const shescape = Shescape({ shell: false });
  * spawn(
  *   "echo",
  *   ["Hello", shescape.escape(userInput)],
@@ -29,7 +39,7 @@ import { checkedToString, toArrayIfNecessary } from "./src/reflection.js";
  * );
  * @example
  * import { spawn } from "node:child_process";
- * const shescape = Shescape();
+ * const shescape = Shescape({ shell: false });
  * spawn(
  *   "echo",
  *   shescape.escapeAll(["Hello", userInput]),
@@ -60,21 +70,25 @@ export class Shescape {
    *
    * @param {object} [options] The escape options.
    * @param {boolean} [options.flagProtection=true] Is flag protection enabled.
-   * @param {boolean} [options.interpolation=true] Is interpolation enabled.
-   * @param {boolean | string} [options.shell] The shell to escape for.
+   * @param {boolean | string} [options.shell=true] The shell to escape for.
+   * @throws {Error} The shell is not supported.
    * @since 2.0.0
    */
   constructor(options = {}) {
     const platform = os.platform();
     const helpers = getHelpersByPlatform({ env: process.env, platform });
 
-    const { flagProtection, interpolation, shellName } = parseOptions(
+    const { flagProtection, shellName } = parseOptions(
       { options, process },
       helpers,
     );
 
+    if (!helpers.isShellSupported(shellName)) {
+      throw new Error(shellError(shellName));
+    }
+
     {
-      const escape = helpers.getEscapeFunction(shellName, { interpolation });
+      const escape = helpers.getEscapeFunction(shellName);
       if (flagProtection) {
         const flagProtect = helpers.getFlagProtectionFunction(shellName);
         this._escape = (arg) => flagProtect(escape(arg));
@@ -135,6 +149,7 @@ export class Shescape {
    * @param {string} arg The argument to quote and escape.
    * @returns {string} The quoted and escaped argument.
    * @throws {TypeError} The argument is not stringable.
+   * @throws {Error} Quoting is not supported with `shell: false`.
    * @since 2.0.0
    */
   quote(arg) {
@@ -152,6 +167,7 @@ export class Shescape {
    * @param {string[]} args The arguments to quote and escape.
    * @returns {string[]} The quoted and escaped arguments.
    * @throws {TypeError} One of the arguments is not stringable.
+   * @throws {Error} Quoting is not supported with `shell: false`.
    * @since 2.0.0
    */
   quoteAll(args) {