+++ test/_arbitraries.js
@@ -189,25 +189,17 @@ export const shescapeArg = () =>
  */
 export const shescapeOptions = () =>
   fc.option(
-    fc.object({
-      key: fc.oneof(
-        fc.string(),
-        fc.constantFrom(
-          "flagProtection",
-          "interpolation",
-          "quoted",
-          "shell",
-          "shellName",
+    fc.record(
+      {
+        flagProtection: fc.boolean(),
+        shell: fc.oneof(
+          fc.boolean(),
+          fc.constantFrom(null, undefined),
+          constants.isWindows ? windowsShell() : unixShell(),
         ),
-      ),
-      values: [
-        fc.boolean(),
-        fc.nat(),
-        fc.string(),
-        unixShell(),
-        windowsShell(),
-      ],
-    }),
+      },
+      { withDeletedKeys: true },
+    ),
     { nil: undefined },
   );
 
@@ -232,7 +224,12 @@ export const unixShell = () => fc.constantFrom(...constants.shellsUnix);
  * @returns {string} Arbitrary non-Unix shell strings.
  */
 export const unsupportedUnixShell = () =>
-  fc.string().filter((v) => !constants.shellsUnix.includes(v));
+  fc.oneof(
+    windowsShell(),
+    fc
+      .stringMatching(/^[0-9A-Za-z]+$/u)
+      .filter((v) => !constants.shellsUnix.includes(v)),
+  );
 
 /**
  * The unsupportedWindowsShell arbitrary generates strings that are not Windows
@@ -241,7 +238,12 @@ export const unsupportedUnixShell = () =>
  * @returns {string} Arbitrary non-Windows shell strings.
  */
 export const unsupportedWindowsShell = () =>
-  fc.string().filter((v) => !constants.shellsWindows.includes(v));
+  fc.oneof(
+    unixShell(),
+    fc
+      .stringMatching(/^[0-9A-Za-z]+\.exe$/u)
+      .filter((v) => !constants.shellsWindows.includes(v)),
+  );
 
 /**
  * The windowsPath arbitrary generates absolute Windows file/folder paths.