+++ test/unit/win/index.test.js
@@ -12,8 +12,9 @@ import sinon from "sinon";
 
 import { arbitrary, constants } from "./_.js";
 
-import * as cmd from "../../../src/win/cmd.js";
 import * as win from "../../../src/win.js";
+import * as cmd from "../../../src/win/cmd.js";
+import * as noShell from "../../../src/win/no-shell.js";
 import * as powershell from "../../../src/win/powershell.js";
 
 const shells = [
@@ -51,30 +52,34 @@ testProp(
   },
 );
 
+test("escape function for no shell", (t) => {
+  const actual = win.getEscapeFunction(null);
+  const expected = noShell.getEscapeFunction();
+  t.is(actual, expected);
+});
+
 for (const { module, shellName } of shells) {
   test(`escape function for ${shellName}`, (t) => {
-    let options = { interpolation: false };
-    t.is(
-      win.getEscapeFunction(shellName, options),
-      module.getEscapeFunction(options),
-    );
-
-    options = { interpolation: true };
-    t.is(
-      win.getEscapeFunction(shellName, options),
-      module.getEscapeFunction(options),
-    );
+    const actual = win.getEscapeFunction(shellName);
+    const expected = module.getEscapeFunction();
+    t.is(actual, expected);
   });
 }
 
 testProp(
   "escape function for unsupported shell",
-  [arbitrary.unsupportedWindowsShell(), fc.boolean()],
-  (t, shellName, interpolation) => {
-    const result = win.getEscapeFunction(shellName, { interpolation });
+  [arbitrary.unsupportedWindowsShell()],
+  (t, shellName) => {
+    const result = win.getEscapeFunction(shellName);
     t.is(result, undefined);
   },
 );
+
+test("quote function for no shell", (t) => {
+  const actual = win.getQuoteFunction(null);
+  const expected = noShell.getQuoteFunction();
+  t.deepEqual(actual, expected);
+});
 
 for (const { module, shellName } of shells) {
   test(`quote function for ${shellName}`, (t) => {
@@ -117,7 +122,7 @@ testProp(
     resolveExecutable.returns(path.join(basePath, shell));
 
     const result = win.getShellName({ env, shell }, { resolveExecutable });
-    t.is(result, constants.binCmd);
+    t.is(result, shell);
   },
 );
 
@@ -142,6 +147,12 @@ testProp(
   },
 );
 
+test("flag protection function for no shell", (t) => {
+  const actual = win.getFlagProtectionFunction(null);
+  const expected = noShell.getFlagProtectionFunction();
+  t.is(actual, expected);
+});
+
 for (const { module, shellName } of shells) {
   test(`flag protection function for ${shellName}`, (t) => {
     const actual = win.getFlagProtectionFunction(shellName);
@@ -156,5 +167,26 @@ testProp(
   (t, shellName) => {
     const result = win.getFlagProtectionFunction(shellName);
     t.is(result, undefined);
+  },
+);
+
+test(`is shell supported, no shell`, (t) => {
+  const actual = win.isShellSupported(null);
+  t.true(actual);
+});
+
+for (const { shellName } of shells) {
+  test(`is shell supported, ${shellName}`, (t) => {
+    const actual = win.isShellSupported(shellName);
+    t.true(actual);
+  });
+}
+
+testProp(
+  "is shell supported for unsupported shell",
+  [arbitrary.unsupportedWindowsShell()],
+  (t, shellName) => {
+    const result = win.isShellSupported(shellName);
+    t.false(result);
   },
 );